+++ server/src/modules/market-making/strategy/dex/alpacastrat.service.ts
@@ -6,10 +6,10 @@ import { StrategyExecutionHistory } from 'src/common/entities/market-making/stra
 import { createStrategyKey } from 'src/common/helpers/strategyKey';
 import { ExchangeInitService } from 'src/modules/infrastructure/exchange-init/exchange-init.service';
 import { CustomLogger } from 'src/modules/infrastructure/logger/logger.service';
-import { ArbitrageStrategyDto } from 'src/modules/market-making/strategy/strategy.dto';
+import { ArbitrageStrategyDto } from 'src/modules/market-making/strategy/config/strategy.dto';
 import { Repository } from 'typeorm';
 
-import { StrategyService } from './strategy.service';
+import { StrategyService } from '../strategy.service';
 
 //This is still in testing: Do not use in production