+++ server/src/modules/market-making/strategy/dex/dex.module.ts
@@ -3,7 +3,7 @@ import { DexAdapterRegistry } from 'src/defi/adapter-registry';
 import { PancakeV3Adapter } from 'src/defi/adapters/pancakeV3.adapter';
 import { UniswapV3Adapter } from 'src/defi/adapters/uniswapV3.adapter';
 
-import { Web3Module } from '../../web3/web3.module';
+import { Web3Module } from '../../../web3/web3.module';
 import { DexVolumeStrategyService } from './dex-volume.strategy.service';
 
 @Module({